separate components + add loading skeletons
